requirements.txt
# Core dependencies
wheel>=0.40.0
torch>=2.0.0
transformers>=4.36.0
datasets>=2.10.0
evaluate>=0.4.0
accelerate>=0.25.0
timm>=0.9.0
Pillow>=10.0.0
qwen_vl_utils>=0.1.0

# Optional performance optimization
flash-attn>=2.5.0

# Utilities
aiohttp>=3.8.0
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
tqdm>=4.65.0

# Additional Hugging Face ecosystem
tokenizers>=0.15.0
safetensors>=0.4.0
huggingface-hub>=0.19.0 